+++ deep_refine/deeppoly/analysis.hh
@@ -24,7 +24,7 @@ Layer_t* get_pred_layer(Network_t* net, Layer_t* curr_layer);
 Expr_t* multiply_expr_with_coeff(Network_t* net, Expr_t* expr, double coeff_inf, double coeff_sup);
 void add_expr(Network_t* net, Expr_t* expr1, Expr_t* expr2);
 bool is_image_verified(Network_t* net);
-bool is_greater(Network_t* net, int index1, int index2);
+bool is_greater(Network_t* net, size_t index1, size_t index2);
 void update_pred_layer_link(Network_t* net, Layer_t* pred_layer);
 unsigned int get_num_thread();
